+++ frontend/openchat-agent/src/typebox.ts
@@ -227,6 +227,11 @@ export const GroupRegisterWebhookArgs = Type.Object({
     avatar: Type.Optional(Type.String()),
 });
 
+export type GroupDeleteHistoryArgs = Static<typeof GroupDeleteHistoryArgs>;
+export const GroupDeleteHistoryArgs = Type.Object({
+    before: Type.BigInt(),
+});
+
 export type GroupSelectedUpdatesArgs = Static<typeof GroupSelectedUpdatesArgs>;
 export const GroupSelectedUpdatesArgs = Type.Object({
     updates_since: Type.BigInt(),
@@ -517,7 +522,12 @@ export const OgPreviewImage = Type.Object({
 });
 
 export type ExchangeId = Static<typeof ExchangeId>;
-export const ExchangeId = Type.Union([Type.Literal("ICPSwap"), Type.Literal("Taco")]);
+export const ExchangeId = Type.Union([
+    Type.Literal("ICPSwap"),
+    Type.Literal("Taco"),
+    Type.Literal("Sonic"),
+    Type.Literal("KongSwap"),
+]);
 
 export type ProposalDecisionStatus = Static<typeof ProposalDecisionStatus>;
 export const ProposalDecisionStatus = Type.Union([
@@ -913,6 +923,7 @@ export const ChatEventType = Type.Union([
     Type.Literal("Frozen"),
     Type.Literal("Unfrozen"),
     Type.Literal("DisappearingMessagesUpdated"),
+    Type.Literal("HistoryDeleted"),
     Type.Literal("MessagePinned"),
     Type.Literal("MessageUnpinned"),
     Type.Literal("MembersJoined"),
@@ -2924,6 +2935,12 @@ export const CommunityDeleteMessagesArgs = Type.Object({
     new_achievement: Type.Boolean(),
 });
 
+export type CommunityDeleteChannelHistoryArgs = Static<typeof CommunityDeleteChannelHistoryArgs>;
+export const CommunityDeleteChannelHistoryArgs = Type.Object({
+    channel_id: ChannelId,
+    before: Type.BigInt(),
+});
+
 export type CommunityRemoveMemberFromChannelArgs = Static<
     typeof CommunityRemoveMemberFromChannelArgs
 >;
@@ -4262,6 +4279,12 @@ export const UserSwapTokensExchangeSwapArgs = Type.Object({
     zero_for_one: Type.Boolean(),
 });
 
+export type UserSwapTokensTacoArgs = Static<typeof UserSwapTokensTacoArgs>;
+export const UserSwapTokensTacoArgs = Type.Object({
+    swap_canister_id: TSPrincipal,
+    treasury_canister_id: TSPrincipal,
+});
+
 export type UserSwapTokensResponse = Static<typeof UserSwapTokensResponse>;
 export const UserSwapTokensResponse = Type.Union([
     Type.Object({
@@ -5581,6 +5604,12 @@ export const MembersResponse = Type.Union([
     }),
 ]);
 
+export type HistoryDeleted = Static<typeof HistoryDeleted>;
+export const HistoryDeleted = Type.Object({
+    before: Type.BigInt(),
+    deleted_by: UserId,
+});
+
 export type RoleChanged = Static<typeof RoleChanged>;
 export const RoleChanged = Type.Object({
     user_ids: Type.Array(UserId),
@@ -6693,12 +6722,6 @@ export const UserSetProfileBackgroundArgs = Type.Object({
     profile_background: Type.Optional(Document),
 });
 
-export type UserSwapTokensTacoArgs = Static<typeof UserSwapTokensTacoArgs>;
-export const UserSwapTokensTacoArgs = Type.Object({
-    swap_canister_id: TSPrincipal,
-    treasury_canister_id: TSPrincipal,
-});
-
 export type UserSwapTokensExchangeArgs = Static<typeof UserSwapTokensExchangeArgs>;
 export const UserSwapTokensExchangeArgs = Type.Union([
     Type.Object({
@@ -6706,6 +6729,12 @@ export const UserSwapTokensExchangeArgs = Type.Union([
     }),
     Type.Object({
         Taco: UserSwapTokensTacoArgs,
+    }),
+    Type.Object({
+        Sonic: UserSwapTokensExchangeSwapArgs,
+    }),
+    Type.Object({
+        KongSwap: UserSwapTokensExchangeSwapArgs,
     }),
 ]);
 
@@ -8047,6 +8076,7 @@ export const LocalUserIndexAccessTokenV2Args = Type.Union([
     Type.Object({
         BotActionByCommand: BotActionByCommandArgs,
     }),
+    Type.Literal("Translate"),
 ]);
 
 export type LocalUserIndexBotSendMessageArgs = Static<typeof LocalUserIndexBotSendMessageArgs>;
@@ -8880,6 +8910,9 @@ export const ChatEvent = Type.Union([
     Type.Object({
         BotUpdated: BotUpdated,
     }),
+    Type.Object({
+        HistoryDeleted: HistoryDeleted,
+    }),
     Type.Literal("FailedToDeserialize"),
 ]);
 
@@ -9120,6 +9153,7 @@ export const GroupCanisterGroupChatSummary = Type.Object({
     membership: Type.Optional(GroupMembership),
     video_call_in_progress: Type.Optional(VideoCall),
     verified: Type.Optional(Type.Boolean()),
+    moderation_flags: Type.Optional(Type.Number()),
 });
 
 export type EventWrapperChatEvent = Static<typeof EventWrapperChatEvent>;
@@ -9170,6 +9204,7 @@ export const GroupCanisterGroupChatSummaryUpdates = Type.Object({
     video_call_in_progress: Type.Optional(OptionUpdateVideoCall),
     any_updates_missed: Type.Optional(Type.Boolean()),
     verified: Type.Optional(Type.Boolean()),
+    moderation_flags: Type.Optional(Type.Number()),
 });
 
 export type DirectChatSummaryUpdates = Static<typeof DirectChatSummaryUpdates>;